fundflow_crm/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1
python-telegram-bot==20.7
docusign-esign==3.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
